% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{assert_bbox_intersects_scotland}
\alias{assert_bbox_intersects_scotland}
\title{Assert Bbox Intersects Scotland}
\usage{
assert_bbox_intersects_scotland(bbox, warn_if_not_within = FALSE)
}
\arguments{
\item{bbox}{A \code{SpatExtent} defining the bbox (or a \code{SpatVector}).}

\item{warn_if_not_within}{A flag to indicate whether to perform the additional check.}
}
\description{
Asserts that a given bbox includes part of Scotland's land surface, meaning the
Recreational Potential can be computed. If it does not, an error is raised.
}
\details{
The test performed is whether the given area \emph{intersects} with the boundaries
of Scotland's land surface, using \link[terra:relate]{terra::relate} with \code{relation = "intersects"}.
This also works when \code{bbox} is a \code{SpatVector} defining a more complex geometry
than a simple bounding box.

It is optional to also test whether any part of the bbox falls outside of the valid
area, and print a warning message if so. This calculation is more expensive so is
switched off by default.
}
