% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_to_bool}
\alias{assert_to_bool}
\title{Assert to Bool}
\usage{
assert_to_bool(func)
}
\arguments{
\item{func}{A function which can error out.}
}
\value{
The wrapped function.
}
\description{
Convert an function that throws errors under certain conditions to one
that instead returns FALSE and prints the error message. If the function
does not encounter an error, TRUE is returned.
}
\details{
This is achieved by wrapping the function execution in a \code{tryCatch} and
capturing any errors or warnings.
}
