% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{assert_valid_bbox}
\alias{assert_valid_bbox}
\title{Assert Valid Bbox}
\usage{
assert_valid_bbox(
  bbox,
  min_area = 10000,
  max_area = 1e+09,
  warn_if_not_within = FALSE
)
}
\arguments{
\item{bbox}{A \code{SpatExtent} defining the bbox.}

\item{min_area}{The minimum allowable area in meters.}

\item{max_area}{The maximum allowable area in meters.}

\item{warn_if_not_within}{A flag to indicate whether to perform the additional check.}
}
\description{
Asserts that a given bounding box (bbox) is valid from the perspective of
calculating Recreational Potential. If it is not, an error is raised.
}
\details{
There are currently two tests:
\itemize{
\item Test 1: Bbox includes part of Scotland's land surface.
\item Test 2: Bbox is a valid size.
}

Test 1 tests whether the given area \emph{intersects} with the boundaries of
Scotland's land surface, using \link[terra:relate]{terra::relate} with \code{relation = "intersects"}.
This also works when \code{bbox} is a \code{SpatVector} defining a more complex geometry
than a simple bounding box. It is optional to also test whether any part of the
bbox falls outside of the valid area, and print a warning message if so.
This calculation is more expensive so is switched off by default.

Test 2 raises an error if the bbox has an area smaller than \code{min_area} or
larger than \code{max_area}.
}
\keyword{internal}
