% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_messages}
\alias{capture_messages}
\title{Capture Messages}
\usage{
capture_messages(func)
}
\arguments{
\item{func}{A function that includes messages.}
}
\value{
The wrapped function.
}
\description{
Modify a function so that, when executed, any messages that would
usually be printed to stdout are instead captured and held in a
variable. The wrapped function returns a pair \verb{(result, message)}
that contains the original result and the captured messages.
}
\keyword{internal}
