% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{check_valid_bbox}
\alias{check_valid_bbox}
\title{Check Valid Bbox}
\usage{
check_valid_bbox(...)
}
\description{
Alternative to \link{assert_valid_bbox} that returns either
\code{TRUE} or an instance of \code{simpleError} depending on whether the conditions
are met or not.
}
\seealso{
\link{assert_valid_bbox} the underlying function.

\link{errors_as_messages} the wrapper.
}
\keyword{internal}
