% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreation.R
\name{compute_component}
\alias{compute_component}
\title{Compute Component}
\usage{
compute_component(
  component,
  persona,
  bbox,
  data_dir = NULL,
  skip_checks = FALSE
)
}
\arguments{
\item{component}{The name of the component to compute (one of "SLSRA", "FIPS_N", "FIPS_I", "Water").}

\item{persona}{A named vector containing the persona scores (for all components).}

\item{bbox}{An optional bounding box for cropping.}

\item{data_dir}{Path to the directory containing the rasters.}

\item{skip_checks}{Pass \code{TRUE} to skip checks of the validity of the inputs.}
}
\value{
A single-layered \code{SpatRaster} which is the contribution to the Recreational Potential from this component.
}
\description{
Compute a single component of the Recreational Potential.
}
\details{
This function computes one \emph{component}, \eqn{p_i}, of the Recreational Potential.
This is is one of
\itemize{
\item SLSRA
\item FIPS_N
\item FIPS_I
\item Water
}

A component is simply a pixel-wise summation of all of the layers/items/features,
\eqn{f_{i,j}}, (most of which are either 1/present or 0/absent),
weighted by the corresponding persona scores, \eqn{s_{i,j}},

\deqn{ p_i(x, y) = \sum_{j=1}^{n_i} s_{i,j} f_{i,j}(x, y) }
}
