% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_production.R
\name{compute_distance_window}
\alias{compute_distance_window}
\title{compute_distance_window}
\usage{
compute_distance_window(
  infile,
  dist_file,
  target_tiles = 20,
  buffer_dist = 10000
)
}
\arguments{
\item{infile}{Path to file containing the input raster.}

\item{dist_file}{Path that the output distance raster will be written to.}

\item{target_tiles}{n of tiles to split the raster into}

\item{buffer_dist}{(in m) buffer to consider when splitting rasters into windows}
}
\description{
Given a spatrast, it splits it into windows using \link{create_overlapping_windows}, computes the distance to the nearest non-\code{NA} value in a raster
using haversine distance, writing straight to \code{outfile}.Recomposes the main raster using the windows using
\link{crop_to_original_extent} and, saves the output and deletes intermediate files.
}
\keyword{internal}
