% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreation.R
\name{compute_fips_n}
\alias{compute_fips_n}
\title{Compute FIPS_N Component}
\usage{
compute_fips_n(...)
}
\arguments{
\item{...}{Parameters passed straight to \link{compute_component}}
}
\description{
Compute the FIPS_N ("Natural Features Impacting Potential Services") component
of the Recreational Potential. This is a convenience wrapper around
\link{compute_component} with the argument \code{component = "FIPS_N"}.
}
