% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_production.R
\name{compute_proximity_rasters}
\alias{compute_proximity_rasters}
\title{Compute Proximity Rasters}
\usage{
compute_proximity_rasters(indir, outdir, splitting)
}
\arguments{
\item{indir}{Path to directory containing the input rasters.}

\item{outdir}{Path to directory in which to write the ouput rasters.}

\item{splitting}{T or F; weather to use the splitting, computing and recomposing approach or calcuate the distance for the whole layer.}
}
\description{
For each layer in the \code{FIPS_I} and \code{Water} rasters, compute the distance
from each pixel to the nearest feature (non-\code{NA} value) using
\link{compute_distance}, and map these distances to the unit
interval using \link{map_distance_to_unit}.
\itemize{
\item \strong{Preceded by:} \link{stack_all}
}
}
\details{
The intermediate rasters, which are the distances before mapping to the
unit interval, are retained in \code{outdir} with \verb{_dist} in the file stem.

\strong{WARNING:} This function requires a significant amount of memory,
and is likely to crash your computer! It is intended to be run on
a high-memory system such as LUMI.
}
\keyword{internal}
