% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreation.R
\name{compute_water}
\alias{compute_water}
\title{Compute FIPS_I Component}
\usage{
compute_water(...)
}
\arguments{
\item{...}{Parameters passed straight to \link{compute_component}}
}
\description{
Compute the Water component of the Recreational Potential. This is a convenience
wrapper around \link{compute_component} with the argument
\code{component = "Water"}.
}
