% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{download_example_data}
\alias{download_example_data}
\title{Download Example Data}
\usage{
download_example_data(dest = NULL, timeout = 60)
}
\arguments{
\item{dest}{Optional non-default path to a directory in which to save the data.}

\item{timeout}{Number of seconds before timing out.}
}
\description{
Downloads some smaller example dara for testing the model.
}
\keyword{internal}
