% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{errors_as_messages}
\alias{errors_as_messages}
\title{Errors as Messages}
\usage{
errors_as_messages(func)
}
\description{
Given a function that may throw an error, e.g. via \code{stop()}, produce
a function that instead prints the error message without crashing.
}
