% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Get Data Dir}
\usage{
get_data_dir()
}
\value{
Path to the data directory.
}
\description{
Construct the path to the directory containing the raster data by default,
i.e. assuming no argument was provided to \link{download_data}.
In the event that the data directory is not found, a warning message will be
printed, but the path will still be returned.
}
