% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_feature_mappings}
\alias{get_feature_mappings}
\title{Get Feature Mappings}
\usage{
get_feature_mappings(config)
}
\arguments{
\item{config}{A loaded configuration.}
}
\value{
A named list containing named lists, with the structure
\{layer_name : \{feature_name : raster_value\}\}
}
\description{
Load mappings from features to raster values from a loaded configuration.
}
\details{
Given a loaded configuration, this function returns a mapping from
features to the corresponding values in the raw raster file. These mappings
are furthermore grouped into datasets.
}
\examples{
library(biodt.recreation)
config <- load_config()
mappings <- get_feature_mappings(config)
}
