% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persona.R
\name{get_preset_persona_file}
\alias{get_preset_persona_file}
\title{Get Preset Persona File}
\usage{
get_preset_persona_file()
}
\value{
Path to the preset persona file.
}
\description{
Get the path to the CSV file containing preset personas.
}
