% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_error}
\alias{is_error}
\title{Is Error}
\usage{
is_error(value)
}
\arguments{
\item{value}{The value to test.}
}
\value{
\code{TRUE} if the value is an error, \code{FALSE} otherwise.
}
\description{
Test if a value is an instance of \code{simpleError}.
}
\keyword{internal}
