% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{is_valid_bbox}
\alias{is_valid_bbox}
\title{Is Valid Bbox}
\usage{
is_valid_bbox(...)
}
\description{
Alternative to \link{assert_valid_bbox} that returns \code{TRUE}
or \code{FALSE} depending on whether the conditions are met or not.
}
\seealso{
\link{assert_valid_bbox} the underlying function.

\link{assert_to_bool} the wrapper.
}
\keyword{internal}
