% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persona.R
\name{list_csv_files}
\alias{list_csv_files}
\title{List CSV Files}
\usage{
list_csv_files(dir)
}
\arguments{
\item{dir}{The directory in which to look.}
}
\value{
A list of names
}
\description{
Returns a list of the names of all CSV files in a directory.
}
\keyword{internal}
