% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_files}
\alias{list_files}
\title{List Files}
\usage{
list_files(dir_, ext, recursive = TRUE)
}
\arguments{
\item{dir_}{The directory in which to search for files.}

\item{ext}{The file extension, e.g. "csv", "tif".}

\item{recursive}{If \code{TRUE}, search subdirectories.}
}
\value{
Named list of files.
}
\description{
Returns a named list of files with a given extension in a given
directory, potentially recursing into subdirectories. The named
list has the structure ( file_stem = file_path ).
}
\keyword{internal}
