% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persona.R
\name{list_personas_in_file}
\alias{list_personas_in_file}
\title{List Personas in File}
\usage{
list_personas_in_file(persona_file)
}
\arguments{
\item{persona_file}{The path to a persona file.}
}
\value{
A list of names.
}
\description{
Returns a list of personas in a given file.
}
\keyword{internal}
