% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_production.R
\name{map_distance_to_unit}
\alias{map_distance_to_unit}
\title{Map distances to the unit interval}
\usage{
map_distance_to_unit(infile, outfile, kappa = 6, alpha = 0.01011)
}
\arguments{
\item{kappa}{A location parameter, of little importance frankly.}

\item{alpha}{Coefficient in the exponent}

\item{x}{A raster}
}
\description{
Uses a logistic function to map positive distances \eqn{d}
to the unit interval \eqn{x \in [0, 1]}.
}
\details{
\deqn{ x = \frac{\kappa + 1}{\kappa + \exp(\alpha d)} }
}
\keyword{internal}
