% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_production.R
\name{one_hot_layer}
\alias{one_hot_layer}
\title{One Hot Layer}
\usage{
one_hot_layer(infile, outfile, feature_mapping)
}
\arguments{
\item{infile}{Path to file containing the categorical raster.}

\item{outfile}{Path that the one-hot raster will be written to.}

\item{feature_mapping}{The mapping from feature names to raster values.}
}
\description{
Convert a single-layered integer-valued raster to a 'one-hot' representation
in which each unique value in the input raster is mapped to a unit vector in
a multi-dimensional output raster, where the number of dimensions in the output
equals the number of unique values in the input.
}
\keyword{internal}
