% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_production.R
\name{quantise_slope}
\alias{quantise_slope}
\title{Quantise Slope}
\usage{
quantise_slope(layer)
}
\arguments{
\item{layer}{A single-layered raster with non-negative values.}
}
\value{
The quantised raster.
}
\description{
Takes a raster of absolute gradients and quantise them into 6 categories
according to their steepness.
}
\details{
More generally, this function takes a non-negative, continous-valued raster
and uses \link[terra:classify]{terra::classify} to create an integer-valued raster with values
\verb{\{1, 2, 3, 4, 5, 6\}} using the following intervals:

\deqn{ x_{out} = \begin{cases}
1 & 0 \leq x_{in} < 1.72 \\
2 & 1.72 \leq x_{in} < 2.86 \\
3 & 2.86 \leq x_{in} < 5.71 \\
4 & 5.71 \leq x_{in} < 11.31 \\
5 & 11.31 \leq x_{in} < 16.7 \\
6 & 16.7 \leq x_{in}
\end{cases}}

The origin of these intervals is TODO: what exactly?
}
\keyword{internal}
