% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persona.R
\name{read_persona_csv}
\alias{read_persona_csv}
\title{Read Persona CSV}
\usage{
read_persona_csv(csv_path)
}
\arguments{
\item{csv_path}{Path to a CSV file containing the persona(s)}
}
\value{
data.frame containing \emph{at least} two columns: the \code{index} column and the persona(s)
}
\description{
Read a CSV file containing one or more personas.
}
\details{
This function essentially calls \code{readr::read_csv} with \code{col_types} set to reflect the
expected columns in a persona file. That is: one \code{character} column for the index,
and \emph{at least} one integer column for the persona scores.
}
\seealso{
\code{\link[=load_persona]{load_persona()}} loads a single persona.
}
\keyword{internal}
