% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_production.R
\name{reproject_all}
\alias{reproject_all}
\title{Reproject All}
\usage{
reproject_all(indir, outdir)
}
\arguments{
\item{indir}{Path to directory containing the input rasters.}

\item{outdir}{Path to directory in which to write the ouput rasters.}
}
\description{
Reproject all \code{.tif} files in a directory, writing the reprojected rasters
to another directory. The file names must be consistent with the features
as defined in the configuration file.
\itemize{
\item \strong{Precedes:} \link{reproject_all}
}
}
\seealso{
\link{reproject_layer}
}
\keyword{internal}
