% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_production.R
\name{reproject_layer}
\alias{reproject_layer}
\title{Reproject Layer}
\usage{
reproject_layer(infile, outfile)
}
\arguments{
\item{infile}{Path to file containing the raster to be reprojected.}

\item{outfile}{Path that the reprojected raster will be written to.}
}
\description{
Reproject a single-layered raster onto a consistent grid using an
interpolation method suitable for integer-valued data. The resulting
raster has the following properties:
}
\details{
\itemize{
\item \strong{Filetype:} Unsigned 8-t integer (INT1U)
\item \strong{CRS:} EPSG:27700
\item \strong{Resolution:} 20m
\item \strong{Domain:} A bounding box enclosing Scotland
}

The raster is assumed to be integer-valued, and the interpolation method
passed to \link[terra:project]{terra::project} is \code{near}. The single exception is the special
case of \code{FIPS_N_Slope.tif}, where the raster is first quantised using
\link{quantise_slope} before being reprojected.
}
\keyword{internal}
