% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{run_app}
\alias{run_app}
\title{Run the Recreational Potential App}
\usage{
run_app(persona_dir = NULL, data_dir = NULL, ...)
}
\arguments{
\item{persona_dir}{Path to a directory containing persona files.}

\item{data_dir}{Path to the directory containing the raster data.}

\item{...}{Additional arguments passed to \link[shiny:runApp]{shiny::runApp}.}
}
\description{
Run the Recreational Potential App
}
