% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persona.R
\name{save_persona}
\alias{save_persona}
\title{Save Persona}
\usage{
save_persona(persona, csv_path, name, overwrite = FALSE)
}
\arguments{
\item{persona}{A named vector of integers representing the persona}

\item{csv_path}{Path to write a csv file, which may already exist}

\item{name}{A name for the persona}

\item{overwrite}{A boolean flag indicated whether it is permissible to overwrite
an existing personas with the same name}
}
\description{
Save a persona by creating or appending to a CSV file.
}
\details{
Saves a persona to a csv file. If the csv file already exists, the persona
will be appended as a new column, unless the provided name matches an existing
column, in which case it will overwrite the existing data.
}
\examples{
library(biodt.recreation)

# Example persona: 'Hard_Recreationalist'
example_persona <- get_example_persona()

# create a temporary file to save
new_persona_file <- tempfile(fileext = ".csv")

# Save a copy of the Hard_Recreationalist persona
save_persona(example_persona, new_persona_file, "Hard_Copy")

# Check the file contents
read_persona_csv(new_persona_file)

# Create a random new person
feature_names <- load_config()[["Name"]]
random_persona <- stats::setNames(
    sample(0:10, size = 87, replace = TRUE),
    feature_names
)

# Append it to the file
save_persona(random_persona, new_persona_file, "Random")

# Check the file contents
read_persona_csv(new_persona_file)

unlink(new_persona_file)
}
