% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_production.R
\name{stack_all}
\alias{stack_all}
\title{Stack All}
\usage{
stack_all(indir, outdir)
}
\arguments{
\item{indir}{Path to directory containing the input rasters.}

\item{outdir}{Path to directory in which to write the ouput rasters.}
}
\description{
Construct four multi-layered rasters for each of the four components
of the Recreational Potential model, by stacking the individual
one-hot rasters.
\itemize{
\item \strong{Preceded by:} \link{one_hot_all}
\item \strong{Precedes:} \link{compute_proximity_rasters}
}
}
\keyword{internal}
