% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timed}
\alias{timed}
\title{Timed}
\usage{
timed(func)
}
\arguments{
\item{func}{The function to be timed.}
}
\value{
The wrapped function.
}
\description{
Wrap a function in a timer. The resulting function does exactly the same
thing, but upon completion it prints a message containing the time in
taken in seconds.
}
\keyword{internal}
