% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_to_bool}
\alias{assert_to_bool}
\title{Assert to Bool}
\usage{
assert_to_bool(func)
}
\arguments{
\item{func}{A function which can error out.}
}
\value{
The wrapped function.
}
\description{
Given a function that asserts that some condition is met by throwing
an error if it is not, e.g. via \code{stop()}, produce a function that
instead returns \code{TRUE} or \code{FALSE} depending on whether the condition
is met, and prints the error message without crashing.
}
\details{
This functionality is superseded by the more general
\link{errors_as_messages}.
}
\keyword{internal}
