% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{download_data}
\alias{download_data}
\title{Download Data}
\usage{
download_data(dest = NULL)
}
\arguments{
\item{dest}{Optional non-default path to a directory in which to save the data.}
}
\description{
Downloads the data required to compute Recreational Potential.
}
