% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persona.R
\name{load_persona}
\alias{load_persona}
\title{Load Persona}
\usage{
load_persona(csv_path, name = NULL)
}
\arguments{
\item{csv_path}{Path to a csv file containing one or more personas}

\item{name}{Name of the persona, which should match a column name}
}
\value{
A named vector of integers representing the persona
}
\description{
Load a single persona from a CSV file.
}
\details{
Loads a single persona from a csv file containing one or more personas.
If the file contains more than one persona (i.e. columns), a name specifying
which personal (column) to load must also be provided.
}
\examples{
library(biodt.recreation)

# Path to a CSV file containing one or more personas
csv_path <- get_preset_persona_file()

# Load a single persona from this file
loaded_persona <- load_persona(csv_path, "Hard_Recreationalist")
}
\seealso{
\code{\link[=read_persona_csv]{read_persona_csv()}} reads the entire CSV file
}
